Move InstructionSimplify to separate file
