+++ src/opt.cpp
@@ -126,7 +126,6 @@
 
 using namespace ispc;
 
-static llvm::Pass *CreateInstructionSimplifyPass();
 static llvm::Pass *CreatePeepholePass();
 
 static llvm::Pass *CreateGatherCoalescePass();
@@ -564,159 +563,6 @@ void ispc::Optimize(llvm::Module *module, int optLevel) {
     }
 #endif
 }
-
-///////////////////////////////////////////////////////////////////////////
-
-/** This simple optimization pass looks for a vector select instruction
-    with an all-on or all-off constant mask, simplifying it to the
-    appropriate operand if so.
-
-    @todo The better thing to do would be to submit a patch to LLVM to get
-    these; they're presumably pretty simple patterns to match.
-*/
-class InstructionSimplifyPass : public llvm::FunctionPass {
-  public:
-    InstructionSimplifyPass() : FunctionPass(ID) {}
-
-    llvm::StringRef getPassName() const { return "Vector Select Optimization"; }
-    bool runOnBasicBlock(llvm::BasicBlock &BB);
-    bool runOnFunction(llvm::Function &F);
-
-    static char ID;
-
-  private:
-    static bool simplifySelect(llvm::SelectInst *selectInst, llvm::BasicBlock::iterator iter);
-    static llvm::Value *simplifyBoolVec(llvm::Value *value);
-    static bool simplifyCall(llvm::CallInst *callInst, llvm::BasicBlock::iterator iter);
-};
-
-char InstructionSimplifyPass::ID = 0;
-
-llvm::Value *InstructionSimplifyPass::simplifyBoolVec(llvm::Value *value) {
-    llvm::TruncInst *trunc = llvm::dyn_cast<llvm::TruncInst>(value);
-    if (trunc != NULL) {
-        // Convert trunc({sext,zext}(i1 vector)) -> (i1 vector)
-        llvm::SExtInst *sext = llvm::dyn_cast<llvm::SExtInst>(value);
-        if (sext && sext->getOperand(0)->getType() == LLVMTypes::Int1VectorType)
-            return sext->getOperand(0);
-
-        llvm::ZExtInst *zext = llvm::dyn_cast<llvm::ZExtInst>(value);
-        if (zext && zext->getOperand(0)->getType() == LLVMTypes::Int1VectorType)
-            return zext->getOperand(0);
-    }
-    /*
-      // This optimization has discernable benefit on the perf
-      // suite on latest LLVM versions.
-      // On 3.4+ (maybe even older), it can result in illegal
-      // operations, so it's being disabled.
-    llvm::ICmpInst *icmp = llvm::dyn_cast<llvm::ICmpInst>(value);
-    if (icmp != NULL) {
-        // icmp(ne, {sext,zext}(foo), zeroinitializer) -> foo
-        if (icmp->getSignedPredicate() == llvm::CmpInst::ICMP_NE) {
-            llvm::Value *op1 = icmp->getOperand(1);
-            if (llvm::isa<llvm::ConstantAggregateZero>(op1)) {
-                llvm::Value *op0 = icmp->getOperand(0);
-                llvm::SExtInst *sext = llvm::dyn_cast<llvm::SExtInst>(op0);
-                if (sext)
-                    return sext->getOperand(0);
-                llvm::ZExtInst *zext = llvm::dyn_cast<llvm::ZExtInst>(op0);
-                if (zext)
-                    return zext->getOperand(0);
-            }
-        }
-
-    }
-    */
-    return NULL;
-}
-
-bool InstructionSimplifyPass::simplifySelect(llvm::SelectInst *selectInst, llvm::BasicBlock::iterator iter) {
-    if (selectInst->getType()->isVectorTy() == false)
-        return false;
-    Assert(selectInst->getOperand(1) != NULL);
-    Assert(selectInst->getOperand(2) != NULL);
-    llvm::Value *factor = selectInst->getOperand(0);
-
-    // Simplify all-on or all-off mask values
-    MaskStatus maskStatus = GetMaskStatusFromValue(factor);
-    llvm::Value *value = NULL;
-    if (maskStatus == MaskStatus::all_on)
-        // Mask all on -> replace with the first select value
-        value = selectInst->getOperand(1);
-    else if (maskStatus == MaskStatus::all_off)
-        // Mask all off -> replace with the second select value
-        value = selectInst->getOperand(2);
-    if (value != NULL) {
-        llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, value);
-        return true;
-    }
-
-    // Sometimes earlier LLVM optimization passes generate unnecessarily
-    // complex expressions for the selection vector, which in turn confuses
-    // the code generators and leads to sub-optimal code (particularly for
-    // 8 and 16-bit masks).  We'll try to simplify them out here so that
-    // the code generator patterns match..
-    if ((factor = simplifyBoolVec(factor)) != NULL) {
-        llvm::Instruction *newSelect = llvm::SelectInst::Create(factor, selectInst->getOperand(1),
-                                                                selectInst->getOperand(2), selectInst->getName());
-        llvm::ReplaceInstWithInst(selectInst, newSelect);
-        return true;
-    }
-
-    return false;
-}
-
-bool InstructionSimplifyPass::simplifyCall(llvm::CallInst *callInst, llvm::BasicBlock::iterator iter) {
-    llvm::Function *calledFunc = callInst->getCalledFunction();
-
-    // Turn a __movmsk call with a compile-time constant vector into the
-    // equivalent scalar value.
-    if (calledFunc == NULL || calledFunc != m->module->getFunction("__movmsk"))
-        return false;
-
-    uint64_t mask;
-    if (GetMaskFromValue(callInst->getArgOperand(0), &mask) == true) {
-        llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, LLVMInt64(mask));
-        return true;
-    }
-    return false;
-}
-
-bool InstructionSimplifyPass::runOnBasicBlock(llvm::BasicBlock &bb) {
-    DEBUG_START_PASS("InstructionSimplify");
-
-    bool modifiedAny = false;
-
-restart:
-    for (llvm::BasicBlock::iterator iter = bb.begin(), e = bb.end(); iter != e; ++iter) {
-        llvm::SelectInst *selectInst = llvm::dyn_cast<llvm::SelectInst>(&*iter);
-        if (selectInst && simplifySelect(selectInst, iter)) {
-            modifiedAny = true;
-            goto restart;
-        }
-        llvm::CallInst *callInst = llvm::dyn_cast<llvm::CallInst>(&*iter);
-        if (callInst && simplifyCall(callInst, iter)) {
-            modifiedAny = true;
-            goto restart;
-        }
-    }
-
-    DEBUG_END_PASS("InstructionSimplify");
-
-    return modifiedAny;
-}
-
-bool InstructionSimplifyPass::runOnFunction(llvm::Function &F) {
-
-    llvm::TimeTraceScope FuncScope("InstructionSimplifyPass::runOnFunction", F.getName());
-    bool modifiedAny = false;
-    for (llvm::BasicBlock &BB : F) {
-        modifiedAny |= runOnBasicBlock(BB);
-    }
-    return modifiedAny;
-}
-
-static llvm::Pass *CreateInstructionSimplifyPass() { return new InstructionSimplifyPass; }
 
 ///////////////////////////////////////////////////////////////////////////
 // GatherCoalescePass
